create common util handleTabs
